+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### 🐞 Fixes
 
 - Attempted fix for the "plugin typescript already exists" error.
+- Fixed an issue where negated glob task `inputs` wouldn't filter the file list when hashing.
 
 #### ⚙️ Internal